+++ src/Middleware/AbstractMiddleware.php
@@ -29,7 +29,7 @@ abstract class AbstractMiddleware implements MiddlewareInterface, MiddlewareLogi
     abstract public function handle(
         \stdClass $json,
         ObjectWrapper $object,
-        PropertyMap $map,
+        PropertyMap $propertyMap,
         JsonMapperInterface $mapper
     ): void;
 }