+++ src/Middleware/Debugger.php
@@ -19,14 +19,14 @@ class Debugger extends AbstractMiddleware
         $this->logger = $logger;
     }
 
-    public function handle(\stdClass $json, ObjectWrapper $object, PropertyMap $map, JsonMapperInterface $mapper): void
+    public function handle(\stdClass $json, ObjectWrapper $object, PropertyMap $propertyMap, JsonMapperInterface $mapper): void
     {
         $this->logger->debug(
             'Current state attributes passed through JsonMapper middleware',
             [
                 'json' => json_encode($json),
                 'object' => get_class($object->getObject()),
-                'propertyMap' => $map->toString()
+                'propertyMap' => $propertyMap->toString()
             ]
         );
     }