+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - New Debugger middleware to help debug the in between middleware
+- Caching support to the DocBlockAnnotations and TypedProperties middleware
 
 ## [0.3.0] - 2020-04-13
 ### Added